mejoras en el codigo

+++ src/main.cpp
@@ -4,7 +4,7 @@
 #include <cstdlib>
 
 int main() {
-    srand(static_cast<unsigned int>(time(0)));
+    srand(time(0));
     testLanzarDados();  // Llamada de prueba
     mostrarMenu();
     return 0;
